Reorganize AdMob tracking unit tests (#3471)
### Checklist
- [x] If applicable, unit tests
- [ ] If applicable, create follow-up issues for `purchases-ios` and
hybrids

### Motivation
Tracking internals in `feature/admob` are organized under `tracking/*`,
but tracking-focused unit tests were still in the root `unit/` package.

### Description
- moves existing tracking-focused unit tests into `unit/tracking`
- updates test packages/imports accordingly
- keeps test behavior and assertions unchanged

### Testing
- tests were reorganized only; no production code changes
- `ReadLints` reports no issues for moved test files

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> **Low Risk**
> Low risk because this PR only changes test package
declarations/imports; no production code or behavior is modified.
> 
> **Overview**
> Reorganizes AdMob tracking-focused unit tests by moving them under the
`com.revenuecat.purchases.admob.tracking` test package (updating package
declarations and removing now-unneeded imports).
> 
> Test logic and assertions remain the same; this is a
structural/namespace cleanup to match the `tracking/*` production
organization.
> 
> <sup>Reviewed by [Cursor Bugbot](https://cursor.com/bugbot) for commit
254b57bb5b68aca7ec4f987d0c21971100e12f74. Bugbot is set up for automated
code reviews on this repo. Configure
[here](https://www.cursor.com/dashboard/bugbot).</sup>
<!-- /CURSOR_SUMMARY -->
