+++ feature/admob/src/test/kotlin/com/revenuecat/purchases/admob/unit/tracking/AdMobPrecisionMappingTest.kt
@@ -1,10 +1,9 @@
 @file:OptIn(ExperimentalPreviewRevenueCatPurchasesAPI::class)
 
-package com.revenuecat.purchases.admob
+package com.revenuecat.purchases.admob.tracking
 
 import com.google.android.gms.ads.AdValue
 import com.revenuecat.purchases.ExperimentalPreviewRevenueCatPurchasesAPI
-import com.revenuecat.purchases.admob.tracking.toAdRevenuePrecision
 import com.revenuecat.purchases.ads.events.types.AdRevenuePrecision
 import org.junit.Assert.assertEquals
 import org.junit.Test
