+++ feature/admob/src/test/kotlin/com/revenuecat/purchases/admob/unit/tracking/TrackIfConfiguredTest.kt
@@ -1,10 +1,9 @@
 @file:OptIn(ExperimentalPreviewRevenueCatPurchasesAPI::class)
 
-package com.revenuecat.purchases.admob
+package com.revenuecat.purchases.admob.tracking
 
 import com.revenuecat.purchases.ExperimentalPreviewRevenueCatPurchasesAPI
 import com.revenuecat.purchases.Purchases
-import com.revenuecat.purchases.admob.tracking.trackIfConfigured
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.After
